Priority Dictionary
Instead of using a priority queue as suggested in the comments, I
decided to use a priority dictionary (the same data structure that is
used in color_by_ordering in the lib/coloring/coloring/ordering.py
file).

I used the priority dictionary instead of a priority queue because the
original data structure was a dictionary with the vertex id number being
the key value. Switching to a priority queue would break the upd
function.

Still need to do some memory profiling with theses changes to see how it
effects memory consumption.